+++ Magic/CHANGELOG.md
@@ -46,6 +46,7 @@
 
  - Re-arrange plugin load order to (possibly) fix issues in recent builds with custom
    integrations (such as external attribute registration)
+ - Fix LibsDisguise integration on old legacy versions of LD
 
 # 8.4.3